chat-bot-algerie-telecom/pipelines/depot/depot_code/requirements.txt
numpy
pyyaml
rank-bm25
google-genai
